+++ lib/src/clojure_lsp/startup.clj
@@ -286,6 +286,13 @@
                                             project-settings
                                             force-settings)
                :classpath-settings classpath-settings))
+      (when-let [otlp-config (and (-> @db* :settings :otlp :enabled)
+                                  (-> @db* :settings :otlp :config))]
+        (logger/configure-otlp logger otlp-config)
+        (logger/info "OTLP configured:" {:client (-> @db* :client-info :name)
+                                         :version (-> @db* :client-info :version)
+                                         :username (System/getProperty "user.name")
+                                         :project project-root-uri}))
       (publish-task-progress producer (:analyzing-project task-list) progress-token)
       (logger/info startup-logger-tag "Analyzing source paths for project root" (str root-path))
       (let [analyze-source-paths-fn (if (= :project-namespaces-only (:project-analysis-type @db*))